kcompressd: add v1 - extra kernel thread for mem compression
the idea behind kcompressd is to have an extra kernel thread taking care
of memory compression like its used by zswap and zram. usually the
memory compression is handled by the kswapd kernel thread and that can
be blocked in its swap related activities when memory needs to be
compressed and putting this compression into its own thread frees up
kswapd during that time for doing other things as both are running in
parallel then.

it looks like this can help quite a bit on low memory systems (like many
of the velvet-os supported ones with around 4gb of ram in many cases or
even less) to avoid system stalls when a lot of memory compression has
to be done due to the limited memory in the hardware.

this is v1 of this patch and that is still in its very early stages, but
due to some limited testing of mine it seems to run stable even under
heavy load. the idea is to add it to the velvet-os kernels in the long
run maybe if it helps the performance and useability and it is here that
interested people can test it in daily usage to observe if it is running
stable in all situations and if it really helps. for now it will not be
included by default in the velvet-os kernels, so one will have to apply
it in own kernel builds and some prepared test kernels will also be
provided for easier testing. it was tested to merge well into v6.14.6
and it seems to run stable with it even under heavy load.

see the head of the patch for where it comes from originally and some
more info about it. one adjustment was taken from the patchwork
discussion to avoid some warnings - this is the difference between the
.patch.org and the .patch file